+++ src/vmm/src/lib.rs
@@ -76,7 +76,6 @@ use error::{Error, Result, UserResult};
 use kernel::cmdline as kernel_cmdline;
 use kernel::loader as kernel_loader;
 use logger::error::LoggerError;
-use logger::LogOption;
 use logger::{AppInfo, Level, Metric, LOGGER, METRICS};
 use seccomp::{BpfProgram, SeccompFilter};
 use utils::eventfd::EventFd;
@@ -678,18 +677,7 @@ impl Vmm {
         })
     }
 
-    #[cfg(target_arch = "x86_64")]
-    fn log_dirty_pages(&mut self) {
-        // If we're logging dirty pages, post the metrics on how many dirty pages there are.
-        if LOGGER.flags() | LogOption::LogDirtyPages as usize > 0 {
-            METRICS.memory.dirty_pages.add(self.get_dirty_page_count());
-        }
-    }
-
     fn write_metrics(&mut self) -> result::Result<(), LoggerError> {
-        // The dirty pages are only available on x86_64.
-        #[cfg(target_arch = "x86_64")]
-        self.log_dirty_pages();
         LOGGER.log_metrics().map(|_| ())
     }
 
@@ -1338,25 +1326,6 @@ impl Vmm {
         }
         // Currently, we never get to return with Ok(EventLoopExitReason::Break) because
         // we just invoke stop() whenever that would happen.
-    }
-
-    // Count the number of pages dirtied since the last call to this function.
-    // Because this is used for metrics, it swallows most errors and simply doesn't count dirty
-    // pages if the KVM operation fails.
-    #[cfg(target_arch = "x86_64")]
-    fn get_dirty_page_count(&mut self) -> usize {
-        let dirty_pages_in_region = |(slot, memory_region): (usize, &vm_memory::MemoryRegion)| {
-            self.vm
-                .fd()
-                .get_dirty_log(slot as u32, memory_region.size())
-                .map(|v| v.iter().map(|page| page.count_ones() as usize).sum())
-                .unwrap_or(0 as usize)
-        };
-
-        self.vm
-            .memory()
-            .map(|ref mem| mem.map_and_fold(0, dirty_pages_in_region, std::ops::Add::add))
-            .unwrap_or(0)
     }
 
     /// Set the guest boot source configuration.
@@ -2775,7 +2744,7 @@ mod tests {
             show_level: true,
             show_log_origin: true,
             #[cfg(target_arch = "x86_64")]
-            options: vec![LogOption::LogDirtyPages],
+            options: vec![],
         };
         // Flushing metrics before initializing logger is not erroneous.
         assert!(vmm.flush_metrics().is_ok());
@@ -2827,14 +2796,6 @@ mod tests {
                 assert!(line.contains("Guest-boot-time ="));
             }
         }
-    }
-
-    #[cfg(target_arch = "x86_64")]
-    #[test]
-    fn test_dirty_page_count() {
-        let mut vmm = create_vmm_object(InstanceState::Uninitialized);
-        assert_eq!(vmm.get_dirty_page_count(), 0);
-        // Booting an actual guest and getting real data is covered by `kvm::tests::run_code_test`.
     }
 
     #[test]